вот!

+++ Aroma.BussinesLogic/Core/Levels/AdminAPI.cs
@@ -138,6 +138,44 @@ namespace Aroma.BussinesLogic.Core.Levels
             return new ResponseAddProduct { Status = true };
         }
 
+        internal ResponseGetProducts GetAllProductsAdmin()
+        {
+            try
+            {
+                using (var db = new ProductContext())
+                {
+                    var products = db.Products.OrderByDescending(p => p.Name).ToList(); // Сортируем продукты по количеству просмотров
+             
+
+                    return new ResponseGetProducts
+                    {
+                        Status = true,
+                        Products = products.Select(p => new Product
+                        {
+                            Id = p.Id,
+                            Name = p.Name,
+                            Price = p.Price,
+                            ProductType = p.ProductType,
+                            Category = p.Category,
+                            Description = p.Description,
+                            ImageUrl = p.ImageUrl,
+                            Quantity = p.QuantityProd,
+                            Discount = p.Discount,
+                            View = p.View,
+                            PriceWithDiscount = p.PriceWithDiscount,
+                        }).ToList(),
+
+                 
+
+                    };
+                }
+            }
+            catch (Exception ex)
+            {
+                // В случае ошибки возвращаем статус false и сообщение об ошибке
+                return new ResponseGetProducts { Status = false, Message = ex.Message };
+            }
+        }
 
         internal ResponseGetProducts GetAllProducts()
         {
@@ -241,7 +279,16 @@ namespace Aroma.BussinesLogic.Core.Levels
             {
                 using (var db = new SupportContext())
                 {
+                    
                     var userToDelete = await db.Users.FindAsync(userId);
+                    if(userToDelete.Level == UserRole.Admin)
+                    {
+                        return new ResponseSupport
+                        {
+                            Status = false,
+                            StatusMessage = "Error"
+                        };
+                    }
                     if (userToDelete != null)
                     {
                         db.Users.Remove(userToDelete);
@@ -284,20 +331,24 @@ namespace Aroma.BussinesLogic.Core.Levels
 
                     if (user != null)
                     {
-                        UserRole role = (UserRole)Enum.Parse(typeof(UserRole), newRole);
-                        if (role == UserRole.Admin)
+                        UserRole NewRole = (UserRole)Enum.Parse(typeof(UserRole), newRole);
+                        if (user.Level == UserRole.Admin)
+                        {
+                            return new ResponseSupport { Status = false, StatusMessage = "Admin role cannot be changed for other admins." };
+                        };
+                        if (NewRole == UserRole.Admin)
                         {
                             user.Level = UserRole.Admin;
                         }
-                        if (role == UserRole.Moderator)
+                        if (NewRole == UserRole.Moderator)
                         {
                             user.Level = UserRole.Moderator;
                         }
-                        if (role == UserRole.User)
+                        if (NewRole == UserRole.User)
                         {
                             user.Level = UserRole.User;
                         }
-                        if (role == UserRole.None)
+                        if (NewRole == UserRole.None)
                         {
                             user.Level = UserRole.None;
                         }
